ci: Build & test workflow (#2)
This PR adds basic CI workflow that automatically builds and tests the app. When you push code or make a PR, it builds the app in Release mode, zips it up with the commit hash as the filename (prerequisite for craft), and runs the unit tests in Debug mode. The zip file preserves all the symlinks and executable bit permissions, so you can download it and run the app locally (since it's currently signed ad-hoc, you need to remove the macOS quarantine flag first if you want to run it). The tests skip the UI tests since those need a display.